requirements.txt
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
qwen-vl-utils>=0.0.8
deepspeed>=0.15.0
datasets>=3.0.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm
pyyaml
einops
flash-attn>=2.5.0
sentencepiece
tiktoken
